逻辑完善

+++ src/main/java/com/woyacy/service/impl/LeaveWordServiceImpl.java
@@ -6,11 +6,15 @@ import com.woyacy.bean.LeaveWordBean;
 import com.woyacy.bean.PageBean;
 import com.woyacy.dao.LeaveWordMapper;
 import com.woyacy.service.LeaveWordService;
+import org.apache.commons.collections4.CollectionUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.Arrays;
 import java.util.List;
+import java.util.Map;
+
+import static org.apache.commons.collections4.CollectionUtils.isNotEmpty;
 
 @Service
 public class LeaveWordServiceImpl implements LeaveWordService {
@@ -25,18 +29,27 @@ public class LeaveWordServiceImpl implements LeaveWordService {
 
     public void deleteLeaveWordById(Long[] lds) {
         List<Long> ids = Arrays.asList(lds);
-        ids.stream().forEach(id ->{
-            leaveWordMapper.deleteByIds(id);
-        });
-
+        if (isNotEmpty(ids)) {
+            ids.stream().forEach(id -> {
+                leaveWordMapper.deleteByIds(id);
+            });
+        }
     }
 
     @Override
-    public void updateLeaveWordByStatus(LeaveWordBean leaveWordBean) {
-        //更改状态为已读
-        leaveWordBean.setStatus(LeaveWordBean.READ_STATUS);
-        //更新状态
-        leaveWordMapper.updateLeaveWordByld(leaveWordBean);
+    public void updateLeaveWordByStatus(Long [] ids) {
+        List<Long> longs = Arrays.asList(ids);
+        if (CollectionUtils.isNotEmpty(longs)){
+            longs.forEach(id->{
+                LeaveWordBean leaveWordBean = new LeaveWordBean();
+                leaveWordBean.setId(id);
+                //更改状态为已读
+                leaveWordBean.setStatus(LeaveWordBean.READ_STATUS);
+                //更新状态
+                leaveWordMapper.updateLeaveWordByld(leaveWordBean);
+            });
+        }
+
     }
 
     @Override
@@ -57,5 +70,15 @@ public class LeaveWordServiceImpl implements LeaveWordService {
     @Override
     public List<LeaveWordBean> selectAllLeaveWord() {
         return leaveWordMapper.selectAllLeaveWord();
+    }
+
+    @Override
+    public List<LeaveWordBean> findLeaveWordByCondition(Map<String, Object> map) {
+        return leaveWordMapper.findLeaveWordByCondition(map);
+    }
+
+    @Override
+    public void deleteLeaveWordById(Long id) {
+        leaveWordMapper.deleteByIds(id);
     }
 }
